nvim, coc.nvim, ccls: exclude -mabi=lp64
ccls/clang seemed to get hung up on this argument when indexing a arm64
build of the Linux kernel. In my case it was the Xilinx 5.4 kernel. The
indication of this happening was a jump in the number of files parsed in
the nvim status line, followed by just hanging on some random file
number.